refactor: move publish items metrics api to dop
