+++ modules/dicehub/provider.go
@@ -23,20 +23,16 @@ import (
 	"github.com/erda-project/erda-infra/base/servicehub"
 	"github.com/erda-project/erda-infra/providers/httpserver"
 	image "github.com/erda-project/erda/modules/dicehub/image/db"
-	"github.com/erda-project/erda/modules/dicehub/metrics"
-	"github.com/erda-project/erda/providers/metrics/query"
 )
 
 type provider struct {
-	Log         logs.Logger
-	QueryClient query.MetricQuery `autowired:"metricq-client"`
-	DB          *gorm.DB          `autowired:"mysql-client"`
-	Router      httpserver.Router `autowired:"http-router"`
-	ImageDB     *image.ImageConfigDB
+	Log     logs.Logger
+	DB      *gorm.DB          `autowired:"mysql-client"`
+	Router  httpserver.Router `autowired:"http-router"`
+	ImageDB *image.ImageConfigDB
 }
 
 func (p *provider) Init(ctx servicehub.Context) error {
-	metrics.Client = p.QueryClient
 	p.ImageDB = &image.ImageConfigDB{DB: p.DB}
 	return nil
 }
