+++ modules/dop/services/apierrors/errors.go
@@ -530,6 +530,10 @@ var (
 	ErrStartCronProjectPipeline       = err("ErrStartCronProjectPipeline", "开启流水线定时失败")
 	ErrEndCronProjectPipeline         = err("ErrEndCronProjectPipeline", "关闭流水线定时失败")
 	ErrListExecHistoryProjectPipeline = err("ErrListExecHistoryProjectPipeline", "列出流水线执行明细失败")
+
+	ErrGetPublishItem    = err("ErrGetPublishItem", "获取发布内容详情失败")
+	ErrSratisticsErrList = err("ErrSratisticsErrList", "获取错误列表失败")
+	ErrCrashRateList     = err("ErrCrashRateList", "获取崩溃率失败")
 )
 
 func err(template, defaultValue string) *errorresp.APIError {
