--- modules/openapi/api/apis/dicehub/dicehub_publish_item_statistics_channels.go
@@ -1,34 +1,0 @@
-// Copyright (c) 2021 Terminus, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package dicehub
-
-import (
-	"github.com/erda-project/erda/apistructs"
-	"github.com/erda-project/erda/modules/openapi/api/apis"
-)
-
-var DICEHUB_PUBLISH_ITEM_CHANNELS = apis.ApiSpec{
-	Path:         "/api/publish-items/<publishItemId>/statistics/channels",
-	BackendPath:  "/api/publish-items/<publishItemId>/statistics/channels",
-	Host:         "dicehub.marathon.l4lb.thisdcos.directory:10000",
-	Scheme:       "http",
-	Method:       "GET",
-	RequestType:  apistructs.PublishItemStatisticsDetailRequest{},
-	ResponseType: apistructs.PublishItemStatisticsDetailData{},
-	IsOpenAPI:    true,
-	CheckLogin:   true,
-	CheckToken:   true,
-	Doc:          `summary: 渠道详情，明细数据`,
-}
