+++ modules/openapi/api/apis/dop/publish_item_err_list.go
@@ -12,16 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package dicehub
+package dop
 
-import "github.com/erda-project/erda/modules/openapi/api/apis"
+import (
+	"github.com/erda-project/erda/apistructs"
+	"github.com/erda-project/erda/modules/openapi/api/apis"
+)
 
-var DICEHUB_PUBLISH_ITEM_METIRCS_COMMON = apis.ApiSpec{
-	Path:        "/api/publish-items/<publishItemId>/metrics/<metricName>",
-	BackendPath: "/api/publish-items/<publishItemId>/metrics/<metricName>",
-	Host:        "dicehub.marathon.l4lb.thisdcos.directory:10000",
-	Scheme:      "http",
-	Method:      "GET",
-	IsOpenAPI:   true,
-	Doc:         `summary: 通用metrcis接口，转发使用`,
+var PUBLISH_ITEM_ERR_LIST = apis.ApiSpec{
+	Path:         "/api/publish-items/<publishItemId>/err/list",
+	BackendPath:  "/api/publish-items/<publishItemId>/err/list",
+	Host:         "dop.marathon.l4lb.thisdcos.directory:9527",
+	Scheme:       "http",
+	Method:       "GET",
+	ResponseType: apistructs.PublishItemStatisticsErrListData{},
+	IsOpenAPI:    true,
+	CheckLogin:   true,
+	CheckToken:   true,
+	Doc:          `summary: 错误报告，错误列表`,
 }
