+++ modules/openapi/api/apis/dop/publish_item_err_rate.go
@@ -12,18 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package dicehub
+package dop
 
 import "github.com/erda-project/erda/modules/openapi/api/apis"
 
-var DICEHUB_PUBLISH_ITEM_CERTIFICATION_GET = apis.ApiSpec{
-	Path:        "/api/publish-items/<publishItemId>/certification",
-	BackendPath: "/api/publish-items/<publishItemId>/certification",
-	Host:        "dicehub.marathon.l4lb.thisdcos.directory:10000",
+var PUBLISH_ITEM_ERR_RATE = apis.ApiSpec{
+	Path:        "/api/publish-items/<publishItemId>/err/rate",
+	BackendPath: "/api/publish-items/<publishItemId>/err/rate",
+	Host:        "dop.marathon.l4lb.thisdcos.directory:9527",
 	Scheme:      "http",
 	Method:      "GET",
 	IsOpenAPI:   true,
 	CheckLogin:  true,
 	CheckToken:  true,
-	Doc:         `summary: 发布内容认证列表`,
+	Doc:         `summary: 错误统计，崩溃率`,
 }
