+++ modules/openapi/api/apis/dop/publish_item_metrics_histogram_common.go
@@ -12,18 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package dicehub
+package dop
 
 import "github.com/erda-project/erda/modules/openapi/api/apis"
 
-var DICEHUB_PUBLISH_ITEM_USERS = apis.ApiSpec{
-	Path:        "/api/publish-items/<publishItemId>/statistics/users",
-	BackendPath: "/api/publish-items/<publishItemId>/statistics/users",
-	Host:        "dicehub.marathon.l4lb.thisdcos.directory:10000",
+var PUBLISH_ITEM_METIRCS_HISTOGRAM_COMMON = apis.ApiSpec{
+	Path:        "/api/publish-items/<publishItemId>/metrics/<metricName>/histogram",
+	BackendPath: "/api/publish-items/<publishItemId>/metrics/<metricName>/histogram",
+	Host:        "dop.marathon.l4lb.thisdcos.directory:9527",
 	Scheme:      "http",
 	Method:      "GET",
 	IsOpenAPI:   true,
-	CheckLogin:  true,
-	CheckToken:  true,
-	Doc:         `summary: 统计大盘，累计用户`,
+	Doc:         `summary: 通用metrcis histogram接口，转发使用`,
 }
