+++ modules/openapi/api/apis/dop/publish_item_statistics_channels.go
@@ -12,22 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package dicehub
+package dop
 
 import (
 	"github.com/erda-project/erda/apistructs"
 	"github.com/erda-project/erda/modules/openapi/api/apis"
 )
 
-var DICEHUB_PUBLISH_ITEM_ERR_LIST = apis.ApiSpec{
-	Path:         "/api/publish-items/<publishItemId>/err/list",
-	BackendPath:  "/api/publish-items/<publishItemId>/err/list",
-	Host:         "dicehub.marathon.l4lb.thisdcos.directory:10000",
+var PUBLISH_ITEM_CHANNELS = apis.ApiSpec{
+	Path:         "/api/publish-items/<publishItemId>/statistics/channels",
+	BackendPath:  "/api/publish-items/<publishItemId>/statistics/channels",
+	Host:         "dop.marathon.l4lb.thisdcos.directory:9527",
 	Scheme:       "http",
 	Method:       "GET",
-	ResponseType: apistructs.PublishItemStatisticsErrListData{},
+	RequestType:  apistructs.PublishItemStatisticsDetailRequest{},
+	ResponseType: apistructs.PublishItemStatisticsDetailData{},
 	IsOpenAPI:    true,
 	CheckLogin:   true,
 	CheckToken:   true,
-	Doc:          `summary: 错误报告，错误列表`,
+	Doc:          `summary: 渠道详情，明细数据`,
 }
