+++ modules/openapi/api/apis/dop/publish_item_statistics_trend.go
@@ -12,22 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package dicehub
+package dop
 
 import (
 	"github.com/erda-project/erda/apistructs"
 	"github.com/erda-project/erda/modules/openapi/api/apis"
 )
 
-var DICEHUB_PUBLISH_ITEM_ERR_TREND = apis.ApiSpec{
-	Path:         "/api/publish-items/<publishItemId>/err/trend",
-	BackendPath:  "/api/publish-items/<publishItemId>/err/trend",
-	Host:         "dicehub.marathon.l4lb.thisdcos.directory:10000",
+var PUBLISH_ITEM_STATISTICS_TREND = apis.ApiSpec{
+	Path:         "/api/publish-items/<publishItemId>/statistics/trend",
+	BackendPath:  "/api/publish-items/<publishItemId>/statistics/trend",
+	Host:         "dop.marathon.l4lb.thisdcos.directory:9527",
 	Scheme:       "http",
 	Method:       "GET",
-	ResponseType: apistructs.PublishItemStatisticsErrTrendData{},
+	ResponseType: apistructs.PublishItemStatisticsTrendData{},
 	IsOpenAPI:    true,
 	CheckLogin:   true,
 	CheckToken:   true,
-	Doc:          `summary: 错误报告，错误趋势`,
+	Doc:          `summary: 统一大盘，整体趋势接口`,
 }
